requirements.txt
fastapi
uvicorn
openai
python-dotenv
pillow
numpy
python-multipart
